+++ src/params/mod.rs
@@ -1,0 +1,8 @@
+use clap::Parser;
+
+#[derive(Parser)]
+pub struct Args {
+    /// Deduplicate icons
+    #[arg(short, long)]
+    pub dedup: bool,
+}
